Rename test account module
Also, create new `testAccounts` object.
