+++ packages/networks/src/index.ts
@@ -1,4 +1,4 @@
 export * from './captureSnapshot.js'
 export * from './createNetwork.js'
-export * from './defaultAccounts.js'
+export * from './testAccounts.js'
 export * from './types.js'
