+++ packages/networks/src/testAccounts.ts
@@ -1,6 +1,6 @@
 import { testingPairs } from '@acala-network/chopsticks-testing'
 
-import type { Keyring } from '@polkadot/keyring'
+import { Keyring } from '@polkadot/keyring'
 import type { KeyringInstance, KeyringPair } from '@polkadot/keyring/types'
 import { cryptoWaitReady } from '@polkadot/util-crypto'
 
@@ -36,3 +36,36 @@ export const defaultAccounts: DefaultAccounts = testingPairs() as DefaultAccount
 export const defaultAccountsSr25519: DefaultAccounts = (await cryptoWaitReady().then(() =>
   testingPairs('sr25519'),
 )) as DefaultAccounts
+
+/**
+ * Type of object containing test accounts without eth keypairs.
+ */
+export type TestAccounts = {
+  alice: KeyringPair
+  bob: KeyringPair
+  charlie: KeyringPair
+  dave: KeyringPair
+  eve: KeyringPair
+  ferdie: KeyringPair
+  keyring: Keyring
+}
+
+/**
+ * Fresh test accounts with seeds different from those used by PJS.
+ *
+ * These accounts use fresh seeds to avoid conflicts with existing dev accounts
+ * that may have been used in testnets.
+ */
+export const testAccounts: TestAccounts = await cryptoWaitReady().then(() => {
+  const keyring = new Keyring({ type: 'sr25519' })
+
+  return {
+    alice: keyring.addFromUri('//fresh_alice'),
+    bob: keyring.addFromUri('//fresh_bob'),
+    charlie: keyring.addFromUri('//fresh_charlie'),
+    dave: keyring.addFromUri('//fresh_dave'),
+    eve: keyring.addFromUri('//fresh_eve'),
+    ferdie: keyring.addFromUri('//fresh_ferdie'),
+    keyring,
+  }
+})
